Rework autotools packaging logic
Rather than having two copies of the autotools files, copy the ones
we're using into the published tarball.

Switch from invoking autotools separately to a single autoreconf.

Don't set SCRYPTVERSION from the directory (that doesn't work any
more now that we're in git); instead, insist on the user setting it.