Switch from (set|long)jmp to sig(set|long)jmp
This allows us to indicate that the signal mask does not need to be
saved and restored, and thus omit some system calls; this is safe
because there is never any opportunity for the signal mask to change
while we bounce around inside the macros registering command-line
options.

In the (admittedly extremely simple) test for getopt, this saves
roughly 40 microseconds of CPU time on my laptop; so this change is
more a matter of tidiness than efficiency.