readpass: Print correct error message if we get EOF
fgets might fail because of EOF - that is, read() returned 0, and
didn't set errno.  In this case, warnp is wrong, because it prints
errno (which will often by something like ENOTTY).

Introduce a new wrapper for warnp used in readpass to detect this
case (with feof) and handle it specially.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>