readpass: Always reissue signals on the failure path
If readpass is interrupted by a signal, fgets fails with EINTR.  It
then prints an error message and goes onto the failure path.

But after doing the cleanup (resetting terminal modes, etc.), we
should reissue the signal.  After all we want a normal program to die
from the signal, rather than calling exit.

A similar argument applies if a signal happens to arrive elsewhere in
readpass.

The main effect of this is that pressing ^C at a password prompt from
a simple utility (like scrypt) now causes the utility's wait status to
report the signal, rather than a failure exit code.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>