Tools: warn about create-*.sh only working on FreeBSD
On Linux and MacOS X, we'd need to use
    shasum -a 256
instead of sha256sum, and even then, it produces a different signing message.
Since this is only intended to be used by cperciva and he runs FreeBSD, I
figure that an error is sufficient for other OSes.  If somebody really needs
to create signed scrypt tarballs on other platforms, it's easy enough to
change this script.