tests: allow multiple exit codes
The previous test suite could only handle a single command in each
??-*.sh scenario file.  This change allows multiple commands to be
tested in each scenario by writing all return values (and valgrind
log files) to separate files.

This change now relies on the setup_check_variables() function and
some "global" variables (prefixed by $s_ or $c_) in order to
minimize ugly boilerplate in the ??-*.sh scenario files.

Note that the awkward
	(
		cmd...
		echo $? > $c_exitfile
	)
construct is used because ${cmd} might not be exiting before we
want the test suite to advance to the next command.  This likely
won't apply to test scenarios in scrypt, but when we apply this
code to spipe, being able to grab the return code of a server
which began 10 shell commands ago (by adding & after the final
parenthesis) is absolutely necessary.  I thought the test suite
(as a whole) would be easier to understand if I used that
construct everywhere, instead of only for certain scenarios.