Travis-CI: use strict compile warnings/errors
This adds code to compile with clang and gcc, since there will be
some difference in their tests.

For the record, the Ubuntu 16.04 "Trusty" image in Travis-CI uses:
- gcc 4.8
- clang 3.5
so the warnings are tailored to those versions.  If/when Travis-CI
updates that build environment, we may need to add other flags
(for example, -Wno-reserved-id-macro to clang).