Check crypto_scrypt return code in test_scrypt
This corrects a case of "tests may fail for the wrong reason": If
crypto_scrypt failed to allocate enough memory, its failure was not
detected prior to this commit; but the fact that it didn't produce
the correct output *was* detected.

Note that the final test case takes 1 GB of memory to compute; small
systems, and even some larger 32-bit systems if they have sufficiently
fragmented address spaces, will not be able to compute it.

Tripped over by:	Tim Bishop
Bug Bounty:	$10 x 2