tests: don't exit if there's no suppressions
grep(1) returns 1 if it fails to find anything, so (in pseudocode)
    grep "remove non-suppression-data from valgrind log file"
was causing a problem on a system without any need of valgrind suppressions.

Unfortunately this means that if grep encounters a real error (exit code 2),
that'll also be ignored... but it didn't seem worth doing something complicated
to accept an error code of 0 or 1, but bail if it was 2.