scryptenc: fix memleak if the input file can't be processed
This is a fix for abfd77eb6962e319a38cfe39c3e7f5b76a134736.

The good news is that the derived keys were never at risk.

The bad news is that because I was so fixated on triple-checking that C->dk
recieved insecure_memzero() in all cases where it contained data, I missed the
fact that in some rare cases we didn't free C.  Namely, this occurs when:
- the input filename successfully responded to open()
- the input file could not be read, or contained the wrong magic header,
  or did not contain the full 96-byte header.

It is safe to call scryptdec_file_cookie_free() on C, even if C->dk hasn't been
initialized, because C->dk is not dynamically allocated.  (We could technically
use free(C) instead of scryptdec_file_cookie_free(), but I thought this fix was
clearer to read.)

The memory leak occured whether or not somebody uses scryptdec_file() or the
newly-exposed scryptdec_file_prep() function.