build: libscrypt_shani -> libcperciva_shani
This is a fix for 390bc93d62ac3107fa7ceaccd9304a93152140ea.

When I added the shani code to scrypt, I copied the definition for
libscrypt_sse2 without realizing that the "scrypt" part was because its code
was lib/crypto/crypto_scrypt_smix_sse2.c.  Since shani code is in
libcperciva/alg/, it should be libcperciva_ instead of libscrypt_.