build: use _XOPEN_SOURCE=700 for strdup() compatibility
The strdup() function was part of XSI in earlier versions of POSIX.  In POSIX
issue 7 (2008), it was moved to Base:
http://pubs.opengroup.org/onlinepubs/9699919799/functions/strdup.html

In theory, the _POSIX_C_SOURCE=200809L should handle it.  However,
glibc 2.12 (2010-05) was the first release to support that; earlier versions
require _XOPEN_SOURCE>=500 to enable it.

We define _XOPEN_SOURCE as well as _POSIX_C_SOURCE in other libcperciva
software.