memlimit: only use RLIMIT_DATA if neither _AS nor _RSS are present
POSIX says of RLIMIT_DATA,

    This is the maximum size of a data segment of the process, in bytes.
    If this limit is exceeded, the malloc() function shall fail with
    errno set to [ENOMEM].
    https://pubs.opengroup.org/onlinepubs/9699919799/functions/getrlimit.html

Many modern malloc() implementations do not allocate space within the
data segment, so this is not relevant to us.  As one NetBSD person put it,

    ... how do you ask malloc to exceed RLIMIT_DATA? Not by passing it a
    large number, or any number, since that doesn't change the size of
    any data segment.
    https://mail-index.netbsd.org/tech-userlevel/2019/07/10/msg011976.html

As such, I think that we should stop considering RLIMIT_DATA when
determining the how much memory to use.  However, not all systems
support RLIMIT_AS and RLIMIT_RSS; so instead of deleting it entirely,
we'll use RLIMIT_DATA as a fallback if neither of the other options are
available.