build: use libscrypt-kdf in main binary
This dynamically links the scrypt binary to libscrypt-kdf if it's being
generated.

(A completely statically linked scrypt is not desirable -- we use libcrypto,
so we need to be able to replace it easily for security fixes.)

It might seem odd that if we're using libscrypt-kdf, we still include a
number of files in scrypt_SOURCES.  That's because we restricted the
symbols in libscrypt-kdf via:

    libscrypt_kdf_la_LDFLAGS=       -version-info 1                 \
        -export-symbols-regex 'crypto_scrypt$$'

This is because the "public API" of libscrypt-kdf is only a single
function -- scrypt_kdf(), which is #define'd as crypto_scrypt -- and we
don't want people starting to rely on any private APIs.