build: fix cpusupport linkage for test_scrypt
Quoted from an email from Colin, lightly edited:

1. main() calls crypto_scrypt()
2. crypto_scrypt() is found in crypto_scrypt.c
3. crypto_scrypt.c includes cpusupport.h
4. cpusupport.h references those symbols.

With optimizations turned on, the compiler can recognize that the
function cpusupport_x86_aesni is static and not used in crypto_scrypt.c
and can thus be removed.  But we can't just assume that the compiler
will optimize things out.

If cpusupport-config.h defines CPUSUPPORT_FOO_BAR, we need to link to
cpusupport_foo_bar.o.