scryptenc_params: return explicit params in scryptdec_{buf,file,prep}
We're replacing the local variables logN, r, p with P->logN, P->r, P->p.
This means that we're returning those values to the caller, via the
structure P.

In the case of scryptdec_file_prep(), we're returning the values *to be
used* (in the future), since that function only prepares for future use.