man: better formatting of conference reference
%O adds a string to the end of a .Rs reference block, meaning that the
date is printed before %O.  However, it's more usual for dates to be at
the end of a reference block.

Unfortunately mdoc doesn't have a "conference" tag.  After testing all
the options, I went with %R for "technical report name", primarily
because then it would be formatted the same as the RFC 7914 block.

It could be argued that conference papers /should/ be formatted
differently than references to RFC documents, but it just looked weird.
If we have multiple references for each type, I'd be happier with
type-specific formatting.