scryptenc_params: add cmdline opts
The bounds for the command-line options in this commit are somewhat
arbitrary.  The actual bounds for the scrypt format are:
   1 <= logN <= 63
   1 <= r
   1 <= p
   (r*p) < 2^30

However, the values at the extremes of those ranges (large and small)
are wildly inappropriate for any practical use.  If somebody wants to
play around with extremely large (or small) values for these parameters,
they can edit main.c or call scryptenc functions manually.

GETOPT_PARENUM_WITHIN_UNSIGNED: many platforms interpret ERANGE as
"result too large", which could be misleading if somebody specified
--logN 8 (where the minimum is 10).  To avoid that problem, I added
special-case handling for that error.  Having done that for --logN, I
figured that I should apply the same to -r and -p as well.