crypto_aesctr_aesni: nonce as a local variable
By itself, this commit has a slight performance penalty (~3.5%).
However, the following commit has a huge speed improvement (more than
double!).  To minimize the following commit, it's useful to deal with
the nonce now.

This commit's penalty could be reduced to ~3% by doing:

-		bufsse = _mm_loadu_si64(stream->pblk + 8);
-		bufsse = _mm_unpacklo_epi64(nonce_be, bufsse);
+		bufsse = _mm_castps_si128(_mm_loadh_pi(
+		    _mm_castsi128_ps(nonce_be),
+		    (const __m64 *)(stream->pblk + 8)));

However, this is much less clear, and it interfers with the gcc
producing good code for the following commit.

For the curious, the alternate code is doing:
    bufsse[63:0] = nonce_be[63:0]
    bufsse[127:64] = (stream->pblk + 8)[63:0]

Unfortunately, the only commands in SSE2 which do this are movhpd and
movhps (_mm_loadh_pd and _mm_loadl_pi).  Since we are dealing with
__m128i, we need to cast to and from floating-point elements.