crypto_aesctr_aesni.c: use a 'do... while' loop
This is to handle clang being fussy; I haven't seen the below problems
with gcc.

Clang needs to know that we will execute the loop at least once.
Otherwise, depending on exactly which compiler flags we pass it and what
future optimizations we use, we face one or more of:
- a compiler warning about a potentially uninitialized variable when we
  try to store the final big-endian block_counter value to
  (stream->pblk + 8) at the end of wholeblocks().
- a 30% speed penalty

The 'for' loop will work with 'assert(num_blocks > 0)' -- but only if we
do not add -DNDEBUG.  That was the final straw which convinced me that
it was worth switching to 'do... while'.