cpusupport: detect a missing _mm_loadu_si64
Old compilers (clang before 2016, gcc before 2019) didn't implement the
_mm_loadu_si64() function.

As for the compiler flags: in some cases, all we need is
    -maes -DBROKEN_MM_LOADU_SI64
but sometimes we might need
    -maes -Wno-cast-align -DBROKEN_MM_LOADU_SI64
or
    -maes -Wno-cast-align -Wno-cast-qual -DBROKEN_MM_LOADU_SI64
or even
    -maes -Wno-cast-align -Wno-cast-qual -Wno-missing-prototypes -DBROKEN_MM_LOADU_SI64

However, the more tests we add to cpusupport.sh, the longer it takes.
I figured that since this is only relevant to old compilers, we might as
well give it the full set of flags, so that newer compilers don't spend
time checking it.