crypto_aes_arm.c: add ARM shims for AESNI intrinsics
These were inspired by the blog posts "Emulating x86 AES instructions on
ARMv8-A [1] and "Porting PuTTY to Windows on Arm" [2].  We're using them
as temporary shims; they will be replaced with more idiomatic native
code in the future.

[1] https://blog.michaelbrase.com/2018/05/08/emulating-x86-aes-intrinsics-on-armv8-a/
[2] https://community.arm.com/developer/tools-software/tools/b/tools-software-ides-blog/posts/porting-putty-to-windows-on-arm

The ARM AES instructions split up AES encryption differently than x86 AESNI.
ARM provides:
	AESE (AES Encrypt)
		a. AddRoundKey
		b. ShiftRows
		c. SubBytes
	AESMC (AES MixColumns)
		a. MixColumns

As a result, we need to play games with the first (or final) XOR.  Both blog
posts have detailed explanations; the best TL;DR is a table from [1]:

	A theoretical 3-round AES

	Round	x86		AES steps		ARM
	-----------------------------------------------------
	1	XOR		AddRoundKey		AESE
		AESENC		SubBytes		"
		"		ShiftRows		"
		"		MixColumns		AESMC
		"		AddRoundKey		AESE
	2	AESENC		SubBytes		"
		"		ShiftRows		"
		"		Mixcolumns		AESMC
		"		AddRoundKey		AESE
	3	AESENCLAST	SubBytes		"
		"		ShiftRows		"
		"		AddRoundKey		XOR

Therefore, to provide shims for _mm_aesenc_si128_si128() and
_mm_aesenclast_si128(), we need to pass a RoundKey of 0 to AESE, then
XOR the result with the desired RoundKey.

The shim for _mm_aeskeygenassist_si128() is more complicated.  There's no
equivalent ARM intrinsic, so we have to implement it ourselves.  Basically, we
use AESE to get the SubBytes operation, then undo the parts of AES that we
don't want: we give it a RoundKey of 0 (rending AddRoundKey to a no-op), and
undo ShiftRows by shuffling bytes back.  The shim in this commit is a slightly
modified version of the one presented in [1].