Increase maximum value of the p cmdline parameter
The upper limit of the value of the `-p` cmdline parameter is 32, which
is relatively low. Currently, hashing with logN=20, r=8 and p=32 takes
around 90s on my slow CPU (Intel i3-7020U).

This is in contrast to the `--logN` parameter, which can be set up to 40,
which corresponds to 1TB of RAM. Additionally, the program will
automatically pick high values of `p` when given high values of the
`-t` (maxtime) parameter.

So allow specifying higher values of the `p` parameter in cmdline.

Signed-off-by: Mateusz Jończyk <mat.jonczyk@o2.pl>