Makefile.am: add libcperciva_cpusupport_detect.la
libcpusupport_crypto_aes.la needs to resolve symbols such as
cpusupport_x86_aesni_detect_1.  In scrypt, that was previously provided
by cpusupport_x86_aesni.o, which was listed explicitly on the
command-line (via Makefile.am's ${crypto_scrypt_files}).

However, in the tarsnap repository, cpusupport_x86_aesni.o was linked
into libtarsnap.a.  When we tried to add libcperciva_crypto_aes.a, that
created a circular dependency:
- libtarsnap.a needed crypto_aes symbols (which were in
  libcperciva_crypto_aes)
- libcperciva_crypto_aes needed cpusupport_x86_aesni_detect_1 (which
  were in libtarsnap.a)

To avoid the circular dependency, we're moving the cpusupport detection
functionality into its own library.  We're also clarifying the order of
libraries in the scrypt_LDADD list.